samples/msgext-link-unfurling/python/requirements.txt
microsoft-agents-activity>=0.6.0
microsoft-agents-hosting-core>=0.6.0
microsoft-agents-hosting-aiohttp>=0.6.0
microsoft-agents-hosting-teams>=0.6.0
microsoft-agents-authentication-msal>=0.6.0
faker